fix: moved mongo repository to mongo folder
