+++ src/adapter/driver/api/controllers/index.ts
@@ -1,5 +1,5 @@
-import { SystemUserRepository } from "@infra/repositories/systemUserRepository";
-import { SystemProdutoRepository } from "@infra/repositories/systemProdutoRepository";
+import { SystemUserRepository } from "@infra/mongo/repositories/systemUserRepository";
+import { SystemProdutoMongoRepository } from "@infra/mongo/repositories/systemProdutoRepository";
 import { SystemUserUseCase } from "@useCases/user";
 import { UserController } from "./userController";
 import { SystemProdutoUseCase } from "@useCases/produto";
@@ -11,7 +11,7 @@ const systemUserUseCase = new SystemUserUseCase(systemUserRepository);
 
 export const userController = new UserController(systemUserUseCase);
 
-const systemProdutoRepository = new SystemProdutoRepository(Produto);
+const systemProdutoRepository = new SystemProdutoMongoRepository(Produto);
 const systemProdutoUseCase = new SystemProdutoUseCase(systemProdutoRepository);
 
 export const produtoController = new ProdutoController(systemProdutoUseCase);
